Agregamos Getmapping en controller

+++ src/main/java/org/bedu/project/model/Producto.java
@@ -23,7 +23,7 @@ public class Producto {
     @Column (length = 100, nullable = false)
     private String marca;
 
-    private BigDecimal precio;
+    private double precio;
     private int stock;
